+++ firmware/apps/color/include/processing.h
@@ -17,8 +17,8 @@
 
 #include "main.h"
 
-extern int processing_get_data(uint16_t color[3], uint16_t *clear,
+extern int processing_get_data(int color[3], int *clear,
                                bool *within_range, int *range_id);
 
 extern int processing_set_color_space(int color_space);
-extern int processing_set_range(int id, bool high, uint16_t color[3]);
+extern int processing_set_range(int id, bool high, int color[3]);
